Add script to list orphans stores
This new script lists stores on two conditions, the user is deleted or
the user has been moved to a different home server but his store is
still hooked. The stores are output as JSON object with possible guessed
owner information, whether the user was deleted and if the store is
empty as that makes it impossible to determine the previous owner.

Unlike kopano-storeadm -O this script uses GetMailBoxTable to retrieve
the stores instead of GetUserStoresTable which is reported to hang on
some kopano-server instances.